ai-interview-platform/backend/requirements.txt
langchain
langchain_google_genai
python-dotenv
fastapi
uvicorn
